Move some non-LaTeX-related functionality from latex_to_unicode.py to normalize_anth.py
